🎨 Add option to use monochrome icons
This commit adds an option to use monochrome icons in the app. This is achieved by adding a new setting in the preferences screen that allows the user to toggle the use of monochrome icons. The setting is enabled by default.

When the setting is enabled, the icons in the food cards are displayed in a monochrome color scheme.